split oversized tool nudge and routing trim test files
